client: Ensure DNS domains pushed via --dhcp-option will not enable split-DNS
In the update OpenVPN 3 Core Library version 3.11, a new configuration
flag has been added to define how --dhcp-option DOMAIN/DOMAIN-SEARCH are
handled.  The default behaviour in the Core library may be different
between platforms.  For Linux, the default is to treat it as split-DNS
(or tunnel DNS scope, in OpenVPN 3 Linux).

This behaviour differs from what prior OpenVPN 3 Linux releases has
done.  Those releases expects these domains to be treated as ordinary
search domains and if split-DNS is wanted, the user will need to
enable split-DNS with the "--dns-scope tunnel" argument when modifying
the configuration profile.

It is easy to go back to the old behaviour. It is just to set the new
configuration flag, dhcpSearchDomainsAsSplitDomains, to 'false' in the
openvpn::ClientAPI::Config object containing VPN configuration profile.

Signed-off-by: David Sommerseth <davids@openvpn.net>