client/netcfg: Restore --dns-setup-disabled functionality
When the --dns handling was added to v24, it did not account for
the --dns-setup-disabled flag in the configuration profile.  This
flag is stored in the NetCfgTunBuilder::disabled_dns_config bool
variable.  The code paths for the --dhcp-option handling had this
check in place.  This resulted the --dns-setup-disabled only
working for --dhcp-option.

When then was upgraded from Core Library version 3.10 to 3.11
happened, the old code paths for --dhcp-option are now parsed via
the same tun_builder_set_dns_options() method.  This resulted in
the --dns-setup-disabled flag not working at all.

This commit restores this check and adds a log event when DNS
setup is skipped

Signed-off-by: David Sommerseth <davids@openvpn.net>