netcfg/resolved: Ensure glib2 params are available on retries
If a background D-Bus method call results in an exception being called
in the proxy->Call(...) in the systemd-resolved D-Bus proxy code, the
parameters sent to the proxy call may have been released by the
g_dbus_proxy_call_sync() call in DBus::Proxy::Client::Call().

This will result in the bgdata->params content being cleared and
released.

The failback logic here is that the background call lambda will
attempt to do another call to the D-Bus service.  But since the
parameters are unreferenced, it may now be NULL.  And this results
in a glib2 library panic, potentially with a segfault crash as the
result.

By providing a GVariant object with an increased reference count to the
proxy->Call(...) method and decrease and release the GVariant buffer
when the background call has completed, this situation is avoided.

Signed-off-by: David Sommerseth <davids@openvpn.net>