common: Properly parse <connection/> blocks
The core-extensions.hpp code didn't properly handle configuration
profiles containing <connection>...</connection> blocks and treated it
as ordinary key/value options.   But it is to be parsed as a container.

Since this is very close to how inline files are handled, we re-use this
code logic.  The inline_file_option() method was renamed to reflect
that it is more related to option tag handling.

Related code comments are also updated to reflect these changes.

Signed-off-by: David Sommerseth <dazo@eurephia.org>