netcfg: Catch Core library exceptions in method_establish()
The OpenVPN 3 Core Library may throw various exceptions depending on
what kind of error situation is occurring.  If this is not caught, the
openvpn3-service-netcfg process will stop (crash).  To avoid this, we
add a larger catch scope when calling the Core library establish()
method.

This also switches to use fmt::format() for error message generation in
both the new and existing exception handling.

Signed-off-by: David Sommerseth <dazo@eurephia.org>