netcfg: Clarify IP address 'prefix' usage
The CIDR references in the current code to network sizes uses the
term 'prefix', which isn't techinically correct.  A better
terminology is 'prefix size', which better describes the use of
this value.

This has been changed in all of the code related to the netcfg
service, including documentation and unit tests.

The net.openvpn.v3.netcfg.NetworkChange signals the
openvpn3-service-netcfg service emits will provide both the 'prefix'
and 'prefix_size' fields, which will have the same value.  This is
to ensure backwards compatibility with existing code.

The 'prefix' field has been deprecated in the documentation and
will be removed in v28 or later.  The openvpn3-service-aws service
has been updated to only depend on 'prefix_size'.

Signed-off-by: David Sommerseth <dazo@eurephia.org>