netcfg: Split up the NetCfgProxy::Network object construction
The NetCfgProxy::Network class has an optional argument to identify if
this is an 'include' or 'exclude' route.  This can cause some unexpected
traps as this code will be refactored a bit, where new arguments to the
constructor will shift the order of some arguments.

To avoid the ambiguity of the API, having an optional argument for
enabling the "exclude path", split the construction up into two static
functions which properly creates the NetCfgProxy::Network object.

Signed-off-by: David Sommerseth <dazo@eurephia.org>