netcfg: Add support for route metric when assing VPN routes
The current implementation ignored the route metric value passed on from
the VPN session/configuration to the Core library functions setting up
network routes.

This extends the related Network objects in the openvpn3-service-netcfg
service integration with a route metric field and passes this value the
whole way.

This has a breaking D-Bus API change between the openvpn3-service-client
and openvpn3-netcfg-service.  If there is an existing
openvpn3-service-netcfg service which has not been updated, the
openvpn3-service-client process will not be able to setup the network
routes.

Signed-off-by: David Sommerseth <dazo@eurephia.org>