netcfg/proxy: Add service version check for D-Bus API compatibility
Since the recent changes in the openvpn3-service-netcfg service modifies
the D-Bus API for adding new routes (networks), openvpn3-service-client
processes would require to connect to an updated openvpn3-service-netcfg
service.

This adds a version check in the D-Bus proxy code for accessing the
netcfg service.  If it detects the service is of version 25 or older, it
will fallback to use the D-Bus API not sending the route metric field.

Signed-off-by: David Sommerseth <dazo@eurephia.org>