python: Deprecate openvpn3.ConfigParser.SanityCheck()
The purpose of this method is now replaced by the
openvpn3.Configuration.Validate() method.  This method is already called
after the VPN profile has been imported to the OpenVPN 3 Configuration
Manager.  This ensures a more consistent profile validation.

Since the ConfigParser parses the configuration profile before the
profile import step, this cannot be easily replaced by a Validate()
method.

The SanityCheck() method is scheduled for removal some time after the
v29 release.

Signed-off-by: David Sommerseth <dazo@eurephia.org>