netcfg/resolved: Fix g_variant_ref assertion warning on session restarts
When a VPN session is restarted, this error may appear in the logs:

    g_variant_ref: assertion 'value != NULL' failed

This happens when a DNS::resolved::Link::BackgroundCall() is called with
nullptr as the params argumment.  This is common when the D-Bus method
don't take any arguments, such as the
org.freedesktop.resolve1.Link.Revert() method.

The g_variant_ref(), g_variant_ref_sink(), g_variant_unref() and related
calls may complain about incorrect values being passed to them if the
argument is a nullptr.

Add a safeguard around these calls so they are not called with a
nullptr.

Codeberg: 67
Report: https://codeberg.org/OpenVPN/openvpn3-linux/issues/67
Signed-off-by: David Sommerseth <dazo@eurephia.org>