netcfg: Make logged metric details more user friendly
Commit a6f36590297f3 (netcfg: Add support for route metric when assing
VPN routes) added the 'metric:' field in the log messages when adding
new routes.  When the metric value is not provided, it defaults to -1.
This may be confusing for users.

This change will instead log '(default)' if the metric value is
negative, to reflect this.  When the metric value is positive,
the provided value will be logged instead.

Signed-off-by: David Sommerseth <dazo@eurephia.org>